Milestone 7 “Ancient, Inc. 2”
Let's return to your job at Ancient, Inc. The report tool (developed as the task for Milestone 5) was a huge success, but too much work for the HR department.

Now they want to delegate the work of congratulating the employees to the departments, but it's still their responsibility to maintain the employee database, so they want to expose report functionality via API!